% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipediapreview-package.R
\docType{package}
\name{wikipediapreview-package}
\alias{wikipediapreview}
\alias{wikipediapreview-package}
\title{wikipediapreview: Wikipedia Preview for R Markdown Documents}
\description{
Facilitates the use of Wikipedia Preview <https://www.mediawiki.org/wiki/Wikipedia_Preview> in R Markdown documents and websites for showing a popup card with a short summary from Wikipedia when a reader hovers over a link to a Wikipedia article (on a desktop) or taps a link (on a mobile device).
}
\author{
\strong{Maintainer}: Mikhail Popov \email{mikhail@wikimedia.org}

Other contributors:
\itemize{
  \item Wikimedia Foundation [copyright holder]
}

}
\keyword{internal}
