% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{wp_init}
\alias{wp_init}
\title{Initialize Wikipedia Preview}
\usage{
wp_init(
  lang = "en",
  use_unpkg = FALSE,
  with_styles = TRUE,
  use_alt_style = FALSE
)
}
\arguments{
\item{lang}{Language code}

\item{use_unpkg}{Whether to link to the JS file on \href{https://unpkg.com/}{unpkg}
rather than embedding the JS code in the document}

\item{with_styles}{Initialize with package-provided CSS styles}

\item{use_alt_style}{Use an alternative built-in style}
}
\description{
Initializes Wikipedia Preview for use in the document
}
